Cast enums to/from bindgen type alias instead of u32
This may fix type errors if the enum representations are platform-dependent.

+++ gdnative-core/src/lib.rs
@@ -174,5 +174,5 @@ pub fn result_from_sys(err: sys::godot_error) -> GodotResult {
         return Ok(());
     }
 
-    Err(unsafe { mem::transmute(err) })
+    Err(unsafe { mem::transmute(err as u32) })
 }
